+++ src/main/java/com/daduo/api/tiktokapi/model/CreditRequest.java
@@ -8,9 +8,9 @@ import lombok.Data;
 @ApiModel(value = "充值币积分请求Json")
 public class CreditRequest {
     @ApiModelProperty(value = "充值币")
-    private Double credit;
+    private Integer credit;
     @ApiModelProperty(value = "用户ID")
     private Long userId;
     @ApiModelProperty(value = "积分数")
-    private Double points;
+    private Integer points;
 }
